edit type of projectCost, moduleCost, otherTaskCost

+++ src/main/java/com/app2/app2t/domain/pjm/OtherTask.java
@@ -20,7 +20,7 @@ public class OtherTask extends BaseEntity {
 
     /**
      */
-    private Integer taskCost;
+    private Double taskCost;
 
     /**
      */
